Fix benchmarks and lints
